joins (mostly?) working. Some column issues remain
