+++ core/src/main/java/tech/tablesaw/joining/DataFrameJoiner.java
@@ -1,24 +1,11 @@
 package tech.tablesaw.joining;
 
 import com.google.common.collect.Streams;
-import com.google.common.primitives.Ints;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import tech.tablesaw.api.*;
 import tech.tablesaw.columns.Column;
-import tech.tablesaw.columns.booleans.BooleanColumnType;
-import tech.tablesaw.columns.dates.DateColumnType;
-import tech.tablesaw.columns.datetimes.DateTimeColumnType;
-import tech.tablesaw.columns.instant.InstantColumnType;
-import tech.tablesaw.columns.numbers.*;
-import tech.tablesaw.columns.strings.StringColumnType;
-import tech.tablesaw.columns.strings.TextColumnType;
-import tech.tablesaw.columns.times.TimeColumnType;
-import tech.tablesaw.index.*;
 import tech.tablesaw.selection.Selection;
 
 /** Implements joins between two or more Tables */
@@ -32,11 +19,14 @@ public class DataFrameJoiner {
     FULL_OUTER
   }
 
+  private static final String LEFT_RECORD_ID_NAME = "_left_record_id_";
+  private static final String RIGHT_RECORD_ID_NAME = "_right_record_id_";
+
   private static final String TABLE_ALIAS = "T";
 
   private final Table table;
   private final String[] joinColumnNames;
-  private final List<Integer> joinColumnIndexes;
+  private int[] leftJoinColumnIndexes;
   private final AtomicInteger joinTableId = new AtomicInteger(2);
 
   /**
@@ -48,7 +38,7 @@ public class DataFrameJoiner {
   public DataFrameJoiner(Table table, String... joinColumnNames) {
     this.table = table;
     this.joinColumnNames = joinColumnNames;
-    this.joinColumnIndexes = getJoinIndexes(table, joinColumnNames);
+    this.leftJoinColumnIndexes = getJoinIndexes(table, joinColumnNames);
   }
 
   /**
@@ -59,8 +49,13 @@ public class DataFrameJoiner {
    * @param columnNames the column names to find indexes of.
    * @return a list of column indexes within the table.
    */
-  private List<Integer> getJoinIndexes(Table table, String[] columnNames) {
-    return Arrays.stream(columnNames).map(table::columnIndex).collect(Collectors.toList());
+  private int[] getJoinIndexes(Table table, String[] columnNames) {
+    int[] results = new int[columnNames.length];
+    for (int i = 0; i < columnNames.length; i++) {
+      String nm = columnNames[i];
+      results[i] = table.columnIndex(nm);
+    }
+    return results;
   }
 
   /**
@@ -198,266 +193,69 @@ public class DataFrameJoiner {
       boolean keepAllJoinKeyColumns,
       String... table2JoinColumnNames) {
 
-    List<Integer> table2JoinColumnIndexes = getJoinIndexes(table2, table2JoinColumnNames);
-    List<Index> table1Indexes = buildIndexesForJoinColumns(joinColumnIndexes, table1);
-    List<Index> table2Indexes = buildIndexesForJoinColumns(table2JoinColumnIndexes, table2);
+    int[] table2JoinColumnIndexes = getJoinIndexes(table2, table2JoinColumnNames);
+
+    table1 = table1.sortOn(leftJoinColumnIndexes);
+    table2 = table2.sortOn(table2JoinColumnIndexes);
 
     Table result = Table.create(table1.name());
+
     // A set of column indexes in the result table that can be ignored. They are duplicate join
     // keys.
-    Set<Integer> resultIgnoreColIndexes =
+    int[] resultIgnoreColIndexes =
         emptyTableFromColumns(
-            result,
-            table1,
-            table2,
-            joinType,
-            allowDuplicates,
-            table2JoinColumnIndexes,
-            keepAllJoinKeyColumns);
+            result, table1, table2, joinType, allowDuplicates, table2JoinColumnIndexes);
 
-    validateIndexes(table1Indexes, table2Indexes);
-    if (table1.rowCount() == 0 && (joinType == JoinType.LEFT_OUTER || joinType == JoinType.INNER)) {
-      // Handle special case of empty table here so it doesn't fall through to the behavior
-      // that adds rows for full outer and right outer joins
-      if (!keepAllJoinKeyColumns) {
-        result.removeColumns(Ints.toArray(resultIgnoreColIndexes));
-      }
-      return result;
+    // add indexes for outer join processing
+    IntColumn indexLeft = IntColumn.indexColumn(LEFT_RECORD_ID_NAME, table1.rowCount(), 0);
+    table1.addColumns(indexLeft);
+    result.addColumns(IntColumn.create(LEFT_RECORD_ID_NAME));
+
+    IntColumn indexRight = IntColumn.indexColumn(RIGHT_RECORD_ID_NAME, table2.rowCount(), 0);
+    table2.addColumns(indexRight);
+    result.addColumns(IntColumn.create(RIGHT_RECORD_ID_NAME));
+
+    validateIndexes(table1, leftJoinColumnIndexes, table2, table2JoinColumnIndexes);
+
+    if (joinType == JoinType.INNER) {
+
+      joinInner(result, table1, table2, table2JoinColumnIndexes, resultIgnoreColIndexes);
+
+    } else if (joinType == JoinType.LEFT_OUTER) {
+
+      joinLeft(result, table1, table2, table2JoinColumnIndexes, resultIgnoreColIndexes);
+
+    } else if (joinType == JoinType.RIGHT_OUTER) {
+
+      joinRight(result, table1, table2, table2JoinColumnIndexes, resultIgnoreColIndexes);
+
+    } else if (joinType == JoinType.FULL_OUTER) {
+      joinFull(result, table1, table2, table2JoinColumnIndexes, resultIgnoreColIndexes);
     }
+    // System.out.println(result);
+    result.removeColumns(LEFT_RECORD_ID_NAME, RIGHT_RECORD_ID_NAME);
 
-    Selection table1DoneRows = Selection.with();
-    Selection table2DoneRows = Selection.with();
-    // use table 2 for row iteration, which can significantly increase performance
-    if (table1.rowCount() > table2.rowCount() && joinType == JoinType.INNER) {
-      for (Row row : table2) {
-        int ri = row.getRowNumber();
-        if (table2DoneRows.contains(ri)) {
-          // Already processed a selection of table1 that contained this row.
-          continue;
-        }
-        Selection table1Rows =
-            createMultiColSelection(
-                table2, ri, table1Indexes, table1.rowCount(), table2JoinColumnIndexes);
-        Selection table2Rows =
-            createMultiColSelection(
-                table2, ri, table2Indexes, table2.rowCount(), table2JoinColumnIndexes);
-        crossProduct(
-            result,
-            table1,
-            table2,
-            table1Rows,
-            table2Rows,
-            resultIgnoreColIndexes,
-            keepAllJoinKeyColumns);
-
-        table2DoneRows = table2DoneRows.or(table2Rows);
-        if (table2DoneRows.size() == table2.rowCount()) {
-          // Processed all the rows in table1 exit early.
-          if (!keepAllJoinKeyColumns) {
-            result.removeColumns(Ints.toArray(resultIgnoreColIndexes));
-          }
-          return result;
-        }
-      }
-    } else {
-      for (Row row : table1) {
-        int ri = row.getRowNumber();
-        if (table1DoneRows.contains(ri)) {
-          // Already processed a selection of table1 that contained this row.
-          continue;
-        }
-        Selection table1Rows =
-            createMultiColSelection(
-                table1, ri, table1Indexes, table1.rowCount(), joinColumnIndexes);
-        Selection table2Rows =
-            createMultiColSelection(
-                table1, ri, table2Indexes, table2.rowCount(), joinColumnIndexes);
-        if ((joinType == JoinType.LEFT_OUTER || joinType == JoinType.FULL_OUTER)
-            && table2Rows.isEmpty()) {
-          withMissingLeftJoin(
-              result, table1, table1Rows, resultIgnoreColIndexes, keepAllJoinKeyColumns);
-        } else {
-          crossProduct(
-              result,
-              table1,
-              table2,
-              table1Rows,
-              table2Rows,
-              resultIgnoreColIndexes,
-              keepAllJoinKeyColumns);
-        }
-        table1DoneRows = table1DoneRows.or(table1Rows);
-        if (joinType == JoinType.FULL_OUTER || joinType == JoinType.RIGHT_OUTER) {
-          // Update done rows in table2 for full Outer.
-          table2DoneRows = table2DoneRows.or(table2Rows);
-        } else if (table1DoneRows.size() == table1.rowCount()) {
-          // Processed all the rows in table1 exit early.
-          if (!keepAllJoinKeyColumns) {
-            result.removeColumns(Ints.toArray(resultIgnoreColIndexes));
-          }
-          return result;
-        }
-      }
-    }
-
-    // Add all rows from table2 that were not handled already.
-    Selection table2Rows = table2DoneRows.flip(0, table2.rowCount());
-    withMissingRight(
-        result,
-        table1.columnCount(),
-        table2,
-        table2Rows,
-        joinType,
-        table2JoinColumnIndexes,
-        resultIgnoreColIndexes,
-        keepAllJoinKeyColumns);
     if (!keepAllJoinKeyColumns) {
-      result.removeColumns(Ints.toArray(resultIgnoreColIndexes));
+      result = result.removeColumns(resultIgnoreColIndexes);
+      System.out.println(result);
+    } else {
+      renameColumnIndexes(result, table2JoinColumnIndexes, resultIgnoreColIndexes);
     }
     return result;
   }
 
-  private void validateIndexes(List<Index> table1Indexes, List<Index> table2Indexes) {
-    if (table1Indexes.size() != table2Indexes.size()) {
-      throw new IllegalArgumentException(
-          "Cannot join using a different number of indices on each table: "
-              + table1Indexes
-              + " and "
-              + table2Indexes);
-    }
-    for (int i = 0; i < table1Indexes.size(); i++) {
-      if (!table1Indexes.get(i).getClass().equals(table2Indexes.get(i).getClass())) {
-        throw new IllegalArgumentException(
-            "Cannot join using different index types: " + table1Indexes + " and " + table2Indexes);
-      }
-    }
-  }
-
-  /** Build a reverse index for every join column in the table. */
-  private List<Index> buildIndexesForJoinColumns(List<Integer> joinColumnIndexes, Table table) {
-    return joinColumnIndexes.stream().map(c -> indexFor(table, c)).collect(Collectors.toList());
-  }
-
-  /** Create a reverse index for a given column. */
-  private Index indexFor(Table table, int colIndex) {
-    ColumnType type = table.column(colIndex).type();
-    if (type instanceof DateColumnType) {
-      return new IntIndex(table.dateColumn(colIndex));
-    } else if (type instanceof DateTimeColumnType) {
-      return new LongIndex(table.dateTimeColumn(colIndex));
-    } else if (type instanceof InstantColumnType) {
-      return new LongIndex(table.instantColumn(colIndex));
-    } else if (type instanceof TimeColumnType) {
-      return new IntIndex(table.timeColumn(colIndex));
-    } else if (type instanceof StringColumnType) {
-      return new StringIndex(table.stringColumn(colIndex));
-    } else if (type instanceof TextColumnType) {
-      return new StringIndex(table.textColumn(colIndex));
-    } else if (type instanceof IntColumnType) {
-      return new IntIndex(table.intColumn(colIndex));
-    } else if (type instanceof LongColumnType) {
-      return new LongIndex(table.longColumn(colIndex));
-    } else if (type instanceof ShortColumnType) {
-      return new ShortIndex(table.shortColumn(colIndex));
-    } else if (type instanceof BooleanColumnType) {
-      return new ByteIndex(table.booleanColumn(colIndex));
-    } else if (type instanceof DoubleColumnType) {
-      return new DoubleIndex(table.doubleColumn(colIndex));
-    } else if (type instanceof FloatColumnType) {
-      return new FloatIndex(table.floatColumn(colIndex));
-    }
-    throw new IllegalArgumentException("Joining attempted on unsupported column type " + type);
-  }
-
   /**
-   * Given a reverse index find a selection of rows that have the same value as the supplied column
-   * does in the given row index.
-   */
-  private Selection selectionForColumn(Column<?> valueColumn, int rowIndex, Index rawIndex) {
-
-    ColumnType type = valueColumn.type();
-    if (type instanceof DateColumnType) {
-      IntIndex index = (IntIndex) rawIndex;
-      int value = ((DateColumn) valueColumn).getIntInternal(rowIndex);
-      return index.get(value);
-    } else if (type instanceof TimeColumnType) {
-      IntIndex index = (IntIndex) rawIndex;
-      int value = ((TimeColumn) valueColumn).getIntInternal(rowIndex);
-      return index.get(value);
-    } else if (type instanceof DateTimeColumnType) {
-      LongIndex index = (LongIndex) rawIndex;
-      long value = ((DateTimeColumn) valueColumn).getLongInternal(rowIndex);
-      return index.get(value);
-    } else if (type instanceof InstantColumnType) {
-      LongIndex index = (LongIndex) rawIndex;
-      long value = ((InstantColumn) valueColumn).getLongInternal(rowIndex);
-      return index.get(value);
-    } else if (type instanceof StringColumnType) {
-      StringIndex index = (StringIndex) rawIndex;
-      String value = ((StringColumn) valueColumn).get(rowIndex);
-      return index.get(value);
-    } else if (type instanceof TextColumnType) {
-      StringIndex index = (StringIndex) rawIndex;
-      String value = ((TextColumn) valueColumn).get(rowIndex);
-      return index.get(value);
-    } else if (type instanceof IntColumnType) {
-      IntIndex index = (IntIndex) rawIndex;
-      int value = ((IntColumn) valueColumn).getInt(rowIndex);
-      return index.get(value);
-    } else if (type instanceof LongColumnType) {
-      LongIndex index = (LongIndex) rawIndex;
-      long value = ((LongColumn) valueColumn).getLong(rowIndex);
-      return index.get(value);
-    } else if (type instanceof ShortColumnType) {
-      ShortIndex index = (ShortIndex) rawIndex;
-      short value = ((ShortColumn) valueColumn).getShort(rowIndex);
-      return index.get(value);
-    } else if (type instanceof BooleanColumnType) {
-      ByteIndex index = (ByteIndex) rawIndex;
-      byte value = ((BooleanColumn) valueColumn).getByte(rowIndex);
-      return index.get(value);
-    } else if (type instanceof DoubleColumnType) {
-      DoubleIndex index = (DoubleIndex) rawIndex;
-      double value = ((DoubleColumn) valueColumn).getDouble(rowIndex);
-      return index.get(value);
-    } else if (type instanceof FloatColumnType) {
-      FloatIndex index = (FloatIndex) rawIndex;
-      float value = ((FloatColumn) valueColumn).getFloat(rowIndex);
-      return index.get(value);
-    } else {
-      throw new IllegalArgumentException(
-          "Joining is supported on numeric, string, and date-like columns. Column "
-              + valueColumn.name()
-              + " is of type "
-              + valueColumn.type());
-    }
-  }
-  /**
-   * Create a big multicolumn selection for all join columns in the given table. Joins two tables.
+   * Renames the column indexes for the second table from Placeholder_X to their original names
    *
-   * @param table the table that used to generate Selection.
-   * @param ri row number of row in table.
-   * @param indexes a reverse index for every join column in the table.
-   * @param selectionSize max size in table .
-   * @param joinColumnIndexes the column index of join key in tables
-   * @return selection created
+   * @param resultIgnoreColIndexes The positions of the secondary join columns
    */
-  private Selection createMultiColSelection(
-      Table table,
-      int ri,
-      List<Index> indexes,
-      int selectionSize,
-      List<Integer> joinColumnIndexes) {
-    Selection multiColSelection = Selection.withRange(0, selectionSize);
-    int i = 0;
-    for (Integer joinColumnIndex : joinColumnIndexes) {
-      Column<?> col = table.column(joinColumnIndex);
-      Selection oneColSelection = selectionForColumn(col, ri, indexes.get(i));
-      // and the selections.
-      multiColSelection = multiColSelection.and(oneColSelection);
-      i++;
+  private void renameColumnIndexes(
+      Table result, int[] secondaryJoinColumnIndexes, int[] resultIgnoreColIndexes) {
+
+    for (int i = 0; i < resultIgnoreColIndexes.length; i++) {
+      // TODO: Implement me, consider Left vs Right and possible renaming to avoid dupe names
+      // Column c = result.column(resultIgnoreColIndexes.)
     }
-    return multiColSelection;
   }
 
   private String newName(String table2Alias, String columnName) {
@@ -679,8 +477,7 @@ public class DataFrameJoiner {
               allowDuplicateColumnNames,
               false,
               joinColumnNames);
-      joinColumnIndexes.clear();
-      joinColumnIndexes.addAll(getJoinIndexes(joined, joinColumnNames));
+      leftJoinColumnIndexes = getJoinIndexes(joined, joinColumnNames);
     }
     return joined;
   }
@@ -721,7 +518,6 @@ public class DataFrameJoiner {
    *     columns are renamed
    * @param col2Names The columns to join on. If a name refers to a double column, the join is
    *     performed after rounding to integers.
-   * @return The resulting table
    */
   public Table rightOuter(Table table2, boolean allowDuplicateColumnNames, String... col2Names) {
     return joinInternal(
@@ -769,19 +565,18 @@ public class DataFrameJoiner {
    * @param table2 the table on the right side of the join.
    * @param joinType the type of join.
    * @param allowDuplicates whether to allow duplicates. If yes rename columns in table2 that have
-   *     the same name as columns in table1 with the exception of join columns in table2 when
+   *     the same name as columns in table1, with the exception of join columns in table2 when
    *     performing a right join.
    * @param table2JoinColumnIndexes the index locations of the table2 join columns.
    * @return A
    */
-  private Set<Integer> emptyTableFromColumns(
+  private int[] emptyTableFromColumns(
       Table destination,
       Table table1,
       Table table2,
       JoinType joinType,
       boolean allowDuplicates,
-      List<Integer> table2JoinColumnIndexes,
-      boolean keepTable2JoinKeyColumns) {
+      int[] table2JoinColumnIndexes) {
 
     Column<?>[] cols =
         Streams.concat(table1.columns().stream(), table2.columns().stream())
@@ -790,26 +585,27 @@ public class DataFrameJoiner {
 
     // For inner join, left join and full outer join mark the join columns in table2 as
     // placeholders.
-    // For right join mark the join columns in table1 as placeholders.
+    // For right join, mark the join columns in table1 as placeholders.
     // Keep track of which join columns are placeholders so they can be ignored.
-    Set<Integer> ignoreColumns = new HashSet<>();
+    int[] ignoreColumns = new int[leftJoinColumnIndexes.length];
+    int i = 0;
+    int ignoreIndex = 0;
     for (int c = 0; c < cols.length; c++) {
       if (joinType == JoinType.RIGHT_OUTER) {
-        if (c < table1.columnCount() && joinColumnIndexes.contains(c)) {
-          if (!keepTable2JoinKeyColumns) {
-            cols[c].setName("Placeholder_" + ignoreColumns.size());
-          }
-          ignoreColumns.add(c);
+        if (c < table1.columnCount() && indexesContainsValue(leftJoinColumnIndexes, c)) {
+          ignoreColumns[ignoreIndex] = c;
+          cols[c].setName("Placeholder_" + ignoreIndex);
+          ignoreIndex++;
         }
-      } else {
-        int table2Index = c - table1.columnCount();
-        if (c >= table1.columnCount() && table2JoinColumnIndexes.contains(table2Index)) {
-          if (!keepTable2JoinKeyColumns) {
-            cols[c].setName("Placeholder_" + ignoreColumns.size());
-          }
-          ignoreColumns.add(c);
+      } else { // JoinType is LEFT, INNER, or FULL
+        int table2Index = c + table1.columnCount();
+        if (indexesContainsValue(table2JoinColumnIndexes, c)) {
+          ignoreColumns[ignoreIndex] = table2Index;
+          cols[table2Index].setName("Placeholder_" + ignoreIndex);
+          ignoreIndex++;
         }
       }
+      i++;
     }
 
     // Rename duplicate columns in second table
@@ -833,114 +629,257 @@ public class DataFrameJoiner {
     return ignoreColumns;
   }
 
-  /**
-   * Creates cross product for the selection of two tables.
-   *
-   * @param destination the destination table.
-   * @param table1 the table on left of join.
-   * @param table2 the table on right of join.
-   * @param table1Rows the selection of rows in table1.
-   * @param table2Rows the selection of rows in table2.
-   * @param ignoreColumns a set of column indexes in the result to ignore. They are redundant join
-   *     columns.
-   */
-  @SuppressWarnings({"rawtypes", "unchecked"})
-  private void crossProduct(
+  private void joinLeft(
       Table destination,
-      Table table1,
-      Table table2,
-      Selection table1Rows,
-      Selection table2Rows,
-      Set<Integer> ignoreColumns,
-      boolean keepTable2JoinKeyColumns) {
-    for (int c = 0; c < table1.columnCount() + table2.columnCount(); c++) {
-      if (!keepTable2JoinKeyColumns && ignoreColumns.contains(c)) {
-        continue;
+      Table left,
+      Table right,
+      int[] rightJoinColumnIndexes,
+      int[] ignoreColumns) {
+
+    joinInner(destination, left, right, rightJoinColumnIndexes, ignoreColumns);
+    Selection unmatched =
+        left.intColumn(LEFT_RECORD_ID_NAME)
+            .isNotIn(destination.intColumn(LEFT_RECORD_ID_NAME).unique());
+    addLeftOnlyValues(destination, left, unmatched);
+  }
+
+  private void addLeftOnlyValues(Table destination, Table left, Selection unmatched) {
+    for (Row leftRow : left.where(unmatched)) {
+      Row destRow = destination.appendRow();
+      for (int c = 0; c < leftRow.columnCount() - 1; c++) {
+        updateDestinationRow(destRow, leftRow, c, c);
       }
-      int table2Index = c - table1.columnCount();
-      for (int r1 : table1Rows) {
-        for (int r2 : table2Rows) {
-          if (c < table1.columnCount()) {
-            Column t1Col = table1.column(c);
-            destination.column(c).append(t1Col, r1);
-          } else {
-            Column t2Col = table2.column(table2Index);
-            destination.column(c).append(t2Col, r2);
+      // update the index column putting it at the end of the destination table
+      updateDestinationRow(destRow, leftRow, destRow.columnCount() - 2, leftRow.columnCount() - 1);
+    }
+  }
+
+  private void addRightOnlyValues(Table destination, Table left, Table right, Selection unmatched) {
+    int leftColumnCount = left.columnCount();
+    for (Row rightRow : right.where(unmatched)) {
+      Row destRow = destination.appendRow();
+      for (int c = 0; c < rightRow.columnCount() - 1; c++) {
+        updateDestinationRow(destRow, rightRow, c + leftColumnCount - 1, c);
+      }
+      // update the index column putting it at the end of the destination table
+      updateDestinationRow(
+          destRow, rightRow, destRow.columnCount() - 1, rightRow.columnCount() - 1);
+    }
+  }
+
+  private void joinRight(
+      Table destination,
+      Table left,
+      Table right,
+      int[] rightJoinColumnIndexes,
+      int[] ignoreColumns) {
+    joinInner(destination, left, right, rightJoinColumnIndexes, ignoreColumns);
+    Selection unmatched =
+        right
+            .intColumn(RIGHT_RECORD_ID_NAME)
+            .isNotIn(destination.intColumn(RIGHT_RECORD_ID_NAME).unique());
+    addRightOnlyValues(destination, left, right, unmatched);
+  }
+
+  private void joinFull(
+      Table destination,
+      Table left,
+      Table right,
+      int[] rightJoinColumnIndexes,
+      int[] ignoreColumns) {
+    joinInner(destination, left, right, rightJoinColumnIndexes, ignoreColumns);
+    Selection unmatchedLeft =
+        left.intColumn(LEFT_RECORD_ID_NAME)
+            .isNotIn(destination.intColumn(LEFT_RECORD_ID_NAME).unique());
+    Selection unmatchedRight =
+        right
+            .intColumn(RIGHT_RECORD_ID_NAME)
+            .isNotIn(destination.intColumn(RIGHT_RECORD_ID_NAME).unique());
+    addLeftOnlyValues(destination, left, unmatchedLeft);
+    addRightOnlyValues(destination, left, right, unmatchedRight);
+  }
+
+  private void joinInner(
+      Table destination,
+      Table left,
+      Table right,
+      int[] rightJoinColumnIndexes,
+      int[] ignoreColumns) {
+
+    Comparator<Row> comparator = getRowComparator(left, rightJoinColumnIndexes);
+
+    Row leftRow = left.row(0);
+    Row rightRow = right.row(0);
+
+    // Marks the position of the first record in right table that matches a specific join value
+    int mark = -1;
+
+    while (leftRow.hasNext() || rightRow.hasNext()) {
+      if (mark == -1) {
+        while (comparator.compare(leftRow, rightRow) < 0 && leftRow.hasNext()) leftRow.next();
+        while (comparator.compare(leftRow, rightRow) > 0 && rightRow.hasNext()) rightRow.next();
+        mark =
+            rightRow
+                .getRowNumber(); // set the position of the first matching record on the right side
+      }
+      if (comparator.compare(leftRow, rightRow) == 0 && (leftRow.hasNext() || rightRow.hasNext())) {
+        addValues(destination, leftRow, rightRow, ignoreColumns);
+        if (rightRow.hasNext()) {
+          rightRow.next();
+        } else {
+          rightRow.at(mark);
+          if (leftRow.hasNext()) {
+            leftRow.next();
           }
+          mark = -1;
         }
+      } else {
+        if (rightRow.hasNext() && leftRow.hasNext()) {
+          rightRow.at(mark);
+          leftRow.next();
+          mark = -1;
+        } else {
+          if (leftRow.hasNext()) leftRow.next();
+          break;
+        }
+      }
+    }
+    // add the last value if you end on a match
+    if (comparator.compare(leftRow, rightRow) == 0) {
+      addValues(destination, leftRow, rightRow, ignoreColumns);
+    }
+  }
+
+  private Comparator<Row> getRowComparator(Table left, int[] rightJoinColumnIndexes) {
+    List<ColumnIndexPair> pairs = createJoinColumnPairs(left, rightJoinColumnIndexes);
+    return SortKey.getChain(SortKey.create(pairs));
+  }
+
+  private List<ColumnIndexPair> createJoinColumnPairs(Table left, int[] rightJoinColumnIndexes) {
+    List<ColumnIndexPair> pairs = new ArrayList<>();
+    for (int i = 0; i < leftJoinColumnIndexes.length; i++) {
+      ColumnIndexPair columnIndexPair =
+          new ColumnIndexPair(
+              left.column(leftJoinColumnIndexes[i]).type(),
+              leftJoinColumnIndexes[i],
+              rightJoinColumnIndexes[i]);
+      pairs.add(columnIndexPair);
+    }
+    return pairs;
+  }
+
+  private void updateDestinationRow(
+      Row destRow, Row sourceRow, int destColumnIndex, int sourceColumnIndex) {
+    ColumnType type = destRow.getColumnType(destColumnIndex);
+    if (type.equals(ColumnType.INTEGER)) {
+      destRow.setInt(destColumnIndex, sourceRow.getInt(sourceColumnIndex));
+    } else if (type.equals(ColumnType.LONG)) {
+      destRow.setLong(destColumnIndex, sourceRow.getLong(sourceColumnIndex));
+    } else if (type.equals(ColumnType.SHORT)) {
+      destRow.setShort(destColumnIndex, sourceRow.getShort(sourceColumnIndex));
+    } else if (type.equals(ColumnType.STRING)) {
+      destRow.setString(destColumnIndex, sourceRow.getString(sourceColumnIndex));
+    } else if (type.equals(ColumnType.LOCAL_DATE)) {
+      destRow.setPackedDate(destColumnIndex, sourceRow.getPackedDate(sourceColumnIndex));
+    } else if (type.equals(ColumnType.LOCAL_TIME)) {
+      destRow.setPackedTime(destColumnIndex, sourceRow.getPackedTime(sourceColumnIndex));
+    } else if (type.equals(ColumnType.LOCAL_DATE_TIME)) {
+      destRow.setPackedDateTime(destColumnIndex, sourceRow.getPackedDateTime(sourceColumnIndex));
+    } else if (type.equals(ColumnType.INSTANT)) {
+      destRow.setPackedInstant(destColumnIndex, sourceRow.getPackedInstant(sourceColumnIndex));
+    } else if (type.equals(ColumnType.DOUBLE)) {
+      destRow.setDouble(destColumnIndex, sourceRow.getDouble(sourceColumnIndex));
+    } else if (type.equals(ColumnType.FLOAT)) {
+      destRow.setFloat(destColumnIndex, sourceRow.getFloat(sourceColumnIndex));
+    } else if (type.equals(ColumnType.BOOLEAN)) {
+      destRow.setBooleanAsByte(destColumnIndex, sourceRow.getBooleanAsByte(sourceColumnIndex));
+    } else if (type.equals(ColumnType.TEXT)) {
+      destRow.setText(destColumnIndex, sourceRow.getText(sourceColumnIndex));
+    }
+  }
+
+  private void addValues(Table destination, Row leftRow, Row rightRow, int[] ignoreColumns) {
+
+    Row destRow = destination.appendRow();
+    // update positionally, but take into account the RECORD_ID COLUMNS at the end of the dest table
+
+    int leftColumnCount = leftRow.columnCount();
+    int rightColumnCount = rightRow.columnCount();
+
+    // update from the left table first (everythint but the RECORD_ID column)
+    for (int destIdx1 = 0; destIdx1 < leftColumnCount - 1; destIdx1++) {
+      updateDestinationRow(destRow, leftRow, destIdx1, destIdx1);
+    }
+
+    // update from the right table (everythint but the RECORD_ID column)
+    for (int destIdx2 = (leftColumnCount - 1);
+        destIdx2 < (leftColumnCount + rightColumnCount) - 2;
+        destIdx2++) {
+      int rightIndex = destIdx2 - (leftColumnCount - 1);
+      updateDestinationRow(destRow, rightRow, destIdx2, rightIndex);
+    }
+
+    // update the RECORD_ID columns
+    updateDestinationRow(destRow, leftRow, destRow.columnCount() - 2, leftColumnCount - 1);
+    updateDestinationRow(destRow, rightRow, destRow.columnCount() - 1, rightColumnCount - 1);
+  }
+
+  private boolean indexesContainsValue(int[] joinColumnIndexes, int columnIndex) {
+    for (int i : joinColumnIndexes) {
+      if (columnIndex == i) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  private void validateIndexes(
+      Table table1, int[] table1Indexes, Table table2, int[] table2Indexes) {
+    if (table1Indexes.length != table2Indexes.length) {
+      throw new IllegalArgumentException(
+          "Cannot join using a different number of indices on each table: "
+              + Arrays.toString(table1Indexes)
+              + " and "
+              + Arrays.toString(table2Indexes));
+    }
+    for (int i = 0; i < table1Indexes.length; i++) {
+      if (!table1
+          .column(table1Indexes[i])
+          .getClass()
+          .equals(table2.column(table2Indexes[i]).getClass())) {
+        throw new IllegalArgumentException(
+            "Cannot join using different index types: "
+                + Arrays.toString(table1Indexes)
+                + " and "
+                + Arrays.toString(table2Indexes));
       }
     }
   }
 
   /**
-   * Adds rows to destination for each row in table1 with the columns from table2 added as missing
-   * values.
+   * Describes two columns that are to be compared in a sort The columns are expected to be
+   * referenced in two separate rows. The values of left and right provide the column index
+   * (position) in each of the two rows.
    */
-  @SuppressWarnings({"rawtypes", "unchecked"})
-  private void withMissingLeftJoin(
-      Table destination,
-      Table table1,
-      Selection table1Rows,
-      Set<Integer> ignoreColumns,
-      boolean keepTable2JoinKeyColumns) {
-    for (int c = 0; c < destination.columnCount(); c++) {
-      if (!keepTable2JoinKeyColumns && ignoreColumns.contains(c)) {
-        continue;
-      }
-      if (c < table1.columnCount()) {
-        Column t1Col = table1.column(c);
-        for (int index : table1Rows) {
-          destination.column(c).append(t1Col, index);
-        }
-      } else {
-        for (int r1 = 0; r1 < table1Rows.size(); r1++) {
-          destination.column(c).appendMissing();
-        }
-      }
-    }
-  }
+  public static class ColumnIndexPair {
+    final ColumnType type;
+    final int left;
+    final int right;
 
-  /**
-   * Adds rows to destination for each row in table2 with the columns from table1 added as missing
-   * values.
-   */
-  @SuppressWarnings({"rawtypes", "unchecked"})
-  private void withMissingRight(
-      Table destination,
-      int table1ColCount,
-      Table table2,
-      Selection table2Rows,
-      JoinType joinType,
-      List<Integer> col2Indexes,
-      Set<Integer> ignoreColumns,
-      boolean keepTable2JoinKeyColumns) {
-
-    // Add index data from table2 into join column positions in table one.
-    if (joinType == JoinType.FULL_OUTER) {
-      for (int i = 0; i < col2Indexes.size(); i++) {
-        Column t2Col = table2.column(col2Indexes.get(i));
-        for (int index : table2Rows) {
-          destination.column(joinColumnIndexes.get(i)).append(t2Col, index);
-        }
-      }
+    public ColumnIndexPair(ColumnType type, int left, int right) {
+      this.type = type;
+      this.left = left;
+      this.right = right;
     }
 
-    for (int c = 0; c < destination.columnCount(); c++) {
-      if (!keepTable2JoinKeyColumns) {
-        if (ignoreColumns.contains(c) || joinColumnIndexes.contains(c)) {
-          continue;
-        }
-      }
-      if (c < table1ColCount) {
-        for (int r1 = 0; r1 < table2Rows.size(); r1++) {
-          destination.column(c).appendMissing();
-        }
-      } else {
-        Column t2Col = table2.column(c - table1ColCount);
-        for (int index : table2Rows) {
-          destination.column(c).append(t2Col, index);
-        }
-      }
+    @Override
+    public String toString() {
+      final StringBuilder sb = new StringBuilder("ColumnIndexPair{");
+      sb.append("type=").append(type);
+      sb.append(", left=").append(left);
+      sb.append(", right=").append(right);
+      sb.append('}');
+      return sb.toString();
     }
   }
 }
